[Finishes #163842045] remove code from init files
