+++ app/V1/views/officeview.py
@@ -1,6 +1,6 @@
 from flask import jsonify, make_response, request
 from app.V1.views import bluprint
-from app.V1.models import PoliticalOfficeModel
+from app.V1.models.appmodels import PoliticalOfficeModel
 
 
 @bluprint.route("/offices", methods= ["GET"] )
